requirements.txt
aws-cdk.aws-cloudformation
aws-cdk.aws-codebuild
aws-cdk.aws-codepipeline
aws-cdk.aws-codepipeline-actions
aws-cdk.aws-ec2
aws-cdk.aws-ecs
aws-cdk.aws-ecs-patterns
aws-cdk.aws-ecr
aws-cdk.aws-ecr-assets
aws-cdk.aws-s3
aws-cdk.aws-s3-assets

aws-cdk.aws-sqs
aws-cdk.aws-ssm

aws-cdk.core
aws-cdk.cx-api
aws-cdk.region-info